% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_present.R
\name{theme_present}
\alias{theme_present}
\title{Set figure theme for presentations}
\usage{
theme_present(base_size = 20, base_family = "sans")
}
\arguments{
\item{base_family}{}
}
\description{
Set figure theme for presentations
}
\examples{

theme_present(base_size = 20, base_family = "sans")

Other fonts are available, though sans font is
the easiest to implement using the following.

theme_present(base_family = "")

Updating font size is accomplished by changing the base_size.

theme_present(base_size = 16)

}
